Add notice about Vite error
The error "Function called outside component initialization" is encountered by many, and its cause always is obscure to the developer. This issue is commonly raised on Stackoverflow, Github, and other platforms, with no success at receiving a concrete solution.

This was also recently mentioned in issue #61, but no initiative was taken.